+++ front/profile.php
@@ -1,12 +1,16 @@
 <?php
 
-include('../../../inc/includes.php');
+if (!defined('GLPI_ROOT')) {
+    require dirname(__DIR__, 3) . '/inc/includes.php';
+}
 
 use GlpiPlugin\Ticketmigration\Menu;
 use GlpiPlugin\Ticketmigration\MigrationProfile;
 use GlpiPlugin\Ticketmigration\ProfileRight;
 
-Session::checkRight(ProfileRight::RIGHT_VIEW_PROFILES, READ);
+if (!ProfileRight::canViewProfiles()) {
+    Html::displayErrorAndDie(__('You do not have permission to perform this action.'));
+}
 
 global $DB;
 $profiles = [];
@@ -23,7 +27,7 @@ Glpi\Application\View\TemplateRenderer::getInstance()->display(
     '@ticketmigration/profile/index.html.twig',
     [
         'profiles' => $profiles,
-        'can_create' => Session::haveRight(ProfileRight::RIGHT_MANAGE_PROFILES, CREATE),
+        'can_create' => ProfileRight::canManageProfiles(CREATE),
         'form_url' => MigrationProfile::getFormURL(),
     ],
 );